+++ src/main/java/io/nagurea/smsupsdk/common/http/HTTPMethod.java
@@ -3,5 +3,6 @@ package io.nagurea.smsupsdk.common.http;
 public enum HTTPMethod {
     POST,
     GET,
-    DELETE
+    DELETE,
+    PUT
 }