+++ RELEASE-NOTES.md
@@ -24,7 +24,7 @@ Initial version with:
 
 ## [0.6.0](https://github.com/kisscool62/io.nagurea.smsupsdk/compare/v0.5.0...v0.6.0) (2022-07-01)
 * Lists > Delete list
-
+* Lists > Clear list
 
 ### Upgrade Steps
 * [NO ACTION REQUIRED]
@@ -34,6 +34,7 @@ Initial version with:
 
 ### New Features
 * Lists > Delete list
+* Lists > Clear list
 
 ### Bug Fixes